+++ microlens-mtl/src/Lens/Micro/Mtl/Internal.hs
@@ -49,6 +49,8 @@ module Lens.Micro.Mtl.Internal
 )
 where
 
+import Lens.Micro
+import Lens.Micro.Internal (( #. ))
 
 import Control.Applicative
 import Control.Monad.Reader as Reader
@@ -64,9 +66,6 @@ import Control.Monad.Trans.Except
 import Control.Monad.Trans.List
 import Control.Monad.Trans.Identity
 import Control.Monad.Trans.Maybe
--- microlens
-import Lens.Micro
-import Lens.Micro.Internal
 
 #if __GLASGOW_HASKELL__ < 710
 import Data.Monoid