+++ microlens/CHANGELOG.md
@@ -24,7 +24,7 @@
 
 # 0.4.9.1
 
-* Reexported `<&>` from `Data.Functor` (on recent versions of `base`).
+* Reexported `<&>` from `Data.Functor` (for recent versions of `base`).
 
 # 0.4.9